+++ src/main/java/dev/dfonline/flint/feature/impl/FlintCommandFeature.java
@@ -5,13 +5,10 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import dev.dfonline.flint.Flint;
 import dev.dfonline.flint.FlintAPI;
 import dev.dfonline.flint.feature.trait.CommandFeature;
-import dev.dfonline.flint.hypercube.PlayerBadge;
-import dev.dfonline.flint.hypercube.PlayerLocation;
-import dev.dfonline.flint.hypercube.PlayerProfile;
 import dev.dfonline.flint.hypercube.Plot;
+import dev.dfonline.flint.util.DebugReportUtil;
 import dev.dfonline.flint.util.FlintSound;
-import dev.dfonline.flint.util.PaletteColor;
-import dev.dfonline.flint.util.message.Message;
+import dev.dfonline.flint.util.message.MessageUtil;
 import dev.dfonline.flint.util.message.impl.CompoundMessage;
 import dev.dfonline.flint.util.message.impl.SoundMessage;
 import dev.dfonline.flint.util.message.impl.prefix.DebugMessage;
@@ -145,10 +142,10 @@ public class FlintCommandFeature implements CommandFeature {
         ).then(literal("test")
                 .then(literal("locate")
                         .executes(context -> {
-                            debugLocate(Flint.getUser().getPlayer().getNameForScoreboard());
+                            debugLocate(Flint.getUser().getPlayer().getGameProfile().name());
                             return 1;
                         })
-                        .then(argument("player", StringArgumentType.greedyString())
+                        .then(argument("player", StringArgumentType.word())
                                 .executes(context -> {
                                     debugLocate(StringArgumentType.getString(context, "player"));
                                     return 1;
@@ -157,10 +154,10 @@ public class FlintCommandFeature implements CommandFeature {
                 )
                 .then(literal("whois")
                         .executes(context -> {
-                            debugWhois(Flint.getUser().getPlayer().getNameForScoreboard());
+                            debugWhois(Flint.getUser().getPlayer().getGameProfile().name());
                             return 1;
                         })
-                        .then(argument("player", StringArgumentType.greedyString())
+                        .then(argument("player", StringArgumentType.word())
                                 .executes(context -> {
                                     debugWhois(StringArgumentType.getString(context, "player"));
                                     return 1;
@@ -248,7 +245,7 @@ public class FlintCommandFeature implements CommandFeature {
         FlintAPI.setUserFetchUserProfileWithWhois(userFetchUserProfileWithWhois);
 
         if (userFetchUserProfileWithWhois) {
-            ModeTrackerFeature.requestCurrentUserProfile();
+            ModeTrackerFeature.ensureCurrentUserProfile();
             Flint.getUser().sendMessage(new SuccessMessage("flint.command.flint.user_profile.enabled"));
             return;
         }
@@ -293,109 +290,17 @@ public class FlintCommandFeature implements CommandFeature {
     }
 
     private static void debugLocate(String player) {
-        LocateFeature.requestLocate(player).thenAccept(locate -> sendMessageOnClientThread(new DebugMessage(formatLocateResult(locate)))).exceptionally(throwable -> {
-            sendMessageOnClientThread(new ErrorMessage("flint.command.flint.test.locate.fail", Component.text(throwable.getMessage())));
+        LocateFeature.requestLocate(player).thenAccept(locate -> MessageUtil.sendOnClientThread(new DebugMessage(DebugReportUtil.formatLocateResult(locate)))).exceptionally(throwable -> {
+            MessageUtil.sendOnClientThread(new ErrorMessage("flint.command.flint.test.locate.fail", Component.text(throwable.getMessage())));
             return null;
         });
     }
 
     private static void debugWhois(String player) {
-        WhoisFeature.requestWhois(player).thenAccept(profile -> sendMessageOnClientThread(new DebugMessage(formatProfileResult(profile)))).exceptionally(throwable -> {
-            sendMessageOnClientThread(new ErrorMessage("flint.command.flint.test.whois.fail", Component.text(throwable.getMessage())));
+        WhoisFeature.requestWhois(player).thenAccept(profile -> MessageUtil.sendOnClientThread(new DebugMessage(DebugReportUtil.formatProfileResult(profile)))).exceptionally(throwable -> {
+            MessageUtil.sendOnClientThread(new ErrorMessage("flint.command.flint.test.whois.fail", Component.text(throwable.getMessage())));
             return null;
         });
-    }
-
-    private static void sendMessageOnClientThread(Message message) {
-        Flint.getClient().execute(() -> {
-            if (Flint.getClient().player != null) {
-                Flint.getUser().sendMessage(message);
-            }
-        });
-    }
-
-    private static Component formatLocateResult(PlayerLocation locate) {
-        Component report = debugReport("Locate Test");
-        report = appendValue(report, "Player", locate.player());
-        report = appendValue(report, "Mode", locate.mode().getName());
-        report = appendValue(report, "Mode Enum", locate.mode().name());
-        report = appendValue(report, "Node", locate.node() == null ? null : locate.node().getName());
-        report = appendValue(report, "Node Enum", locate.node() == null ? null : locate.node().name());
-        report = appendValue(report, "Node Id", locate.nodeId());
-
-        Plot plot = locate.plot();
-        if (plot == null) {
-            return appendValue(report, "Plot", "none");
-        }
-
-        report = appendSection(report, "Plot");
-        report = appendValue(report, "Plot Id", plot.getId());
-        report = appendValue(report, "Name", plot.getName().getString());
-        report = appendValue(report, "Handle", plot.getHandle());
-        report = appendValue(report, "Whitelisted", plot.isWhitelisted());
-        report = appendValue(report, "Owner", plot.getOwner());
-        report = appendValue(report, "Owner Profile Loaded", plot.getOwnerProfile() != null);
-        report = appendValue(report, "Owner Ranks", plot.getOwnerRanks().toReadableString());
-        report = appendValue(report, "Dev Origin", plot.getDevOrigin());
-        report = appendValue(report, "Detected Size", plot.getDetectedSize());
-        report = appendValue(report, "Assumed Size", plot.getSize());
-        report = appendValue(report, "Has Underground", plot.hasUnderground());
-        report = appendValue(report, "Code Bounds", plot.getCodeBoundsString());
-        return report;
-    }
-
-    private static Component formatProfileResult(PlayerProfile profile) {
-        Component report = debugReport("Whois Test");
-        report = appendValue(report, "Username", profile.userName());
-        report = appendValue(report, "Pronouns", profile.pronouns());
-        report = appendValue(report, "Joined", profile.joined());
-        report = appendValue(report, "About", profile.about());
-        report = appendValue(report, "Ranks", profile.ranks().toReadableString());
-        report = appendValue(report, "Primary Rank", profile.ranks().getPrimaryRankName());
-        report = appendValue(report, "Rank Levels", profile.ranks().toLevelString());
-
-        if (profile.badges().isEmpty()) {
-            return appendValue(report, "Badges", "none");
-        }
-
-        report = appendSection(report, "Badges");
-        for (int i = 0; i < profile.badges().size(); i++) {
-            PlayerBadge badge = profile.badges().get(i);
-            report = appendValue(report, "Badge " + (i + 1), "text " + readable(badge.text()) + ", name " + readable(badge.name()) + ", description " + readable(badge.description()));
-        }
-        return report;
-    }
-
-    private static Component debugReport(String title) {
-        return Component.text(title, PaletteColor.PINK_LIGHT);
-    }
-
-    private static Component appendSection(Component report, String section) {
-        return report.append(Component.newline())
-                .append(Component.text(section + ":", PaletteColor.PINK_LIGHT));
-    }
-
-    private static Component appendValue(Component report, String key, Object value) {
-        return report.append(Component.newline())
-                .append(Component.text(key, PaletteColor.PINK_LIGHT))
-                .append(Component.text(": ", PaletteColor.GRAY_DARK))
-                .append(Component.text(readable(value), PaletteColor.WHITE));
-    }
-
-    private static String readable(Object value) {
-        if (value == null) {
-            return "unknown";
-        }
-
-        if (value instanceof String string && string.isBlank()) {
-            return "none";
-        }
-
-        return value.toString();
-    }
-
-    private static String readable(String value) {
-        return readable((Object) value);
     }
 
 }
